+++ src/main/java/org/ioteatime/meonghanyangserver/filter/ExceptionHandlerFilter.java
@@ -1,0 +1,42 @@
+package org.ioteatime.meonghanyangserver.filter;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import lombok.extern.slf4j.Slf4j;
+import org.ioteatime.meonghanyangserver.common.api.Api;
+import org.ioteatime.meonghanyangserver.common.exception.ApiExceptionImpl;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+@Slf4j
+@Component
+public class ExceptionHandlerFilter extends OncePerRequestFilter {
+    @Override
+    protected void doFilterInternal(
+            HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
+            throws ServletException, IOException {
+        try {
+            filterChain.doFilter(request, response);
+        } catch (ApiExceptionImpl exception) {
+            exceptionHandle(response, exception);
+        }
+    }
+
+    // jwtRequest filter exception handle
+    private void exceptionHandle(HttpServletResponse response, ApiExceptionImpl exception)
+            throws IOException {
+        response.setStatus(exception.getHttpStatus().value());
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+
+        Api<Object> apiResponse = Api.fail(exception);
+        ObjectMapper objectMapper = new ObjectMapper();
+        String jsonResponse = objectMapper.writeValueAsString(apiResponse);
+
+        response.getWriter().write(jsonResponse);
+    }
+}
